Adding some missing tags based on descriptions
I took the time to add some of the missing tags based on the descriptions. I haven't actually verified these by testing out the langs, I only read the descriptions of each lang and added whatever is mentioned as tags especially for the langs that don't have any tags yet.