Add higher z-index to repo icons to make sure they remain usable
This change adds a higher z-index to repo icons. This is meant to address when a lang entry has a too long title or too long creator's name, which often leads to the icons getting pushed to a separate line and to overlap with the other text spans. The overlap can currently cause the icons to be unclickable. The higher z-index fixes t his.