An attempt to move inactive projects into a separate section
I made an attempt to be able to mark projects as inactive, including a short description of what I think makes most sense as inactive (which is no commits for more than two years, or website dead with no commits in recent months), and reworked the website template to hopefully show these inactive projects separately so that active projects get shown more prominently. The inactive list is way less detailed so that it remains easier to maintain, since it needs a separate section in the template handling all the options. A project can be easily moved back to active by simply removing the inactive entry again.

I don't really know if this will work without major bugs on the live website. But I guess the best test would be to simply merge it and try it out, or is there any other way to test this in advance?